Shell not working, Cargo run --example works
I modified main.rs of every example and updated files in /src to the new design. I also
updated svg_generation.rs to allow cargo run --example to successfully
compile and run to generate corresponding SVGs. The only thing is that
the shell is not working for me and I could not view them in the local
server hence to see if there's bug in the SVGs.